refactor(BookVo):把price的类型改为double

+++ src/main/java/tyut/selab/bookservice/vo/BookVo.java
@@ -135,7 +135,7 @@ public class BookVo {
     /**
      * 价格
      */
-    public void setPrice(Integer price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
